Update mustache.version in source code after bumping npm version
With the goal of streamlining our release process, we hook into npm'
"version" lifecycle. In practise that means whenever we bump the version
number with npm, e.g:

```
$ npm version patch
```

We'll execute a script that reads the updated version from
`package.json` and puts that into the source code (`mustache.js`).

With that in place, we're basically ready to publish the package to
npm.